Server/xrops/server/pyproject.toml
[tool.poetry]
name = "kuvis-server"
version = "0.1.0"
description = "KUVIS server"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
fastapi = "^0.78.0"
pytest = "^7.1.2"
uvicorn = "^0.18.2"
numpy = "^1.23.0"
openslide-python = "^1.2.0"
matplotlib = "^3.5.2"
Pillow = "^9.2.0"
scikit-image = "^0.19.3"
scikit-learn = "^1.1.1"
scipy = "^1.8.1"
Jinja2 = "^3.1.2"
pyvips = "^2.2.1"
opencv-python = "^4.6.0"
pydantic = "^1.9.1"
pymongo = "^4.1.1"
zarr = "^2.12.0"
importlib = "^1.0.4"
tifffile = "^2022.5.4"
blosc = "^1.10.6"
empatches = "^0.1.3"
imgviz = "^1.5.0"
wsi-preprocessing = {git = "https://github.com/lucasrla/wsi-preprocessing.git"}
celery = "^5.2.7"
PyQt5 = "^5.15.7"
redis = "^4.3.4"
rq = "^1.10.1"
cryptography = "^37.0.4"
requests = "^2.28.1"
asgiref = "^3.5.2"
celery-pool-asyncio = "^0.2.0"
easydict = "^1.9"
tabulate = "^0.8.10"
opencv-python-headless = "^4.6.0"
pandas = "^1.5.0"
wsi-tile-cleanup = {git = "https://github.com/lucasrla/wsi-tile-cleanup"}
torch = { version = "^1.12.0+cu113", source = "https://download.pytorch.org/whl/torch_stable.html" }
torchvision = {version = "^0.13.0+cu113", source = "https://download.pytorch.org/whl/torch_stable.html", optional=true}
albumentations = "1.3.0"
segmentation_models_pytorch = "0.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
